refactor: 카드 정보 타입(CardInfo) 정의 추가 및 경로 수정
